requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
requests>=2.31.0
plotly>=5.17.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
duckduckgo-search>=3.9.0
newspaper3k>=0.2.8
lxml_html_clean>=0.1.0
scipy>=1.11.0
tqdm>=4.66.0
finance-datareader>=0.9.50
beautifulsoup4>=4.12.0
pykrx>=1.0.20
FinanceDataReader>=1.9.8
